Add optional --start and --end flags to specify a time range to get cloud-run logs from
