+++ src/commands/lambda/__tests__/flare.test.ts
@@ -34,7 +34,6 @@ import {
   getTags,
   getUniqueFileNames,
   summarizeConfig,
-  validateStartEndFlags,
 } from '../flare'
 import * as flareModule from '../flare'
 import {getAWSCredentials, getLambdaFunctionConfig, maskConfig} from '../functions/commons'
@@ -296,44 +295,6 @@ describe('lambda flare', () => {
       expect(code).toBe(0)
       const output = context.stdout.toString()
       expect(output).toMatchSnapshot()
-    })
-  })
-
-  describe('validateStartEndFlags', () => {
-    it('returns [undefined, undefined] when start and end flags are not specified', () => {
-      const errorMessages: string[] = []
-      const res = validateStartEndFlags(undefined, undefined)
-      expect(res).toEqual([undefined, undefined])
-      expect(errorMessages).toEqual([])
-    })
-
-    it('throws error when start is specified but end is not specified', () => {
-      expect(() => validateStartEndFlags('123', undefined)).toThrowErrorMatchingSnapshot()
-    })
-
-    it('throws error when end is specified but start is not specified', () => {
-      expect(() => validateStartEndFlags(undefined, '123')).toThrowErrorMatchingSnapshot()
-    })
-
-    it('throws error when start is invalid', () => {
-      expect(() => validateStartEndFlags('123abc', '200')).toThrowErrorMatchingSnapshot()
-    })
-
-    it('throws error when end is invalid', () => {
-      expect(() => validateStartEndFlags('100', '234abc')).toThrowErrorMatchingSnapshot()
-    })
-
-    it('throws error when start is not before the end time', () => {
-      expect(() => validateStartEndFlags('200', '100')).toThrowErrorMatchingSnapshot()
-    })
-
-    it('sets end time to current time if end time is too large', () => {
-      const now = Date.now()
-      const res = validateStartEndFlags('0', '9999999999999')
-      expect(res).not.toBeUndefined()
-      const [start, end] = res
-      expect(start).toBe(0)
-      expect(end).toEqual(now)
     })
   })
